profile: Delete "begin" and "end" fields from struct trace_block
Actually, they are not being used at all. Originally, they were
supposed to be used at trace_block_enter and trace_block_leave to
calculate time and so on, but it's easier to implement and easier to
understand the code when trace_block structure has all required
fields by itself (count, time, etc).