mk: initfs: Remove explicit commands for creating parent directory
The necessary behavior is handled through $(common_prereqs) that
contains a secondarily-expanded order-only '| $(@D)/.' prerequisite.

This should fix weird and floating bug reproduced during parallel
builds, when (presumably) an initfs target rule was itself applied for
creating its parent directory, inheriting the target-specific
$(src_file):

    cp -r -T $(src_file) $@    # $@ is $(ROOTFS_DIR)/dst_file

Which depends on (order-only) | $(ROOTFS_DIR)/.

    cp -r -T $(src_file) $@    # $@ is now $(ROOTFS_DIR)/.

Resulting in the following error:

    cp: cannot overwrite directory .../rootfs//. with non-directory

An indirect reason of this behavior is a bogus double-slash ('//') as a
separator that made Make choose the $(ROOTFS_DIR)/% rule instead of the
proper $(ROOTFS_DIR)/. one. This is fixed in build-gen.