cmd: testing: fix incorrect parsing of command line arguments
The implementation currently uses optind to get follow-on parameter
of an option, but the problem with this is when getopt is returned,
optind contains the index of the next argv argument for a subsequent
call to getopt(). This commit replaces it with optarg: when getopt
encounters an option it's follow-on parameter is found in optarg.

Fixes: #1725

Signed-off-by: Mohana Datta Yelugoti <ymdatta.work@gmail.com>