cmd: testing: block_dev_test: add number of blocks to be tested
This patch adds functionality for 'block_dev_test' command to
specify the number of blocks to be tested. This when combined
with start block option, we can narrow down what blocks needs
to be tested. It also adds information about this option to be
displayed in the help message. This fixes part of #1709.

Signed-off-by: Mohana Datta Yelugoti <ymdatta.work@gmail.com>