cmd: testing: block_dev_test: add multiple read/writes blocks option
This patch adds functionality for 'block_dev_test' command to
specify multiple read/write blocks in a single iteration. It also
adds information about this option to be displayed in the help
message. Default value of multiple blocks is set as 4, in case if
the argument is not given by the user but the flag is provided.
This fixes part of #1709.

Signed-off-by: Mohana Datta Yelugoti <ymdatta.work@gmail.com>