cmd: testing: block_dev_test: multiple read/writes for flash device
This patch adds functionality for 'block_dev_test' command to
specify multiple read/write blocks for a flash device in a single
iteration. It also fixes out of block range bug of the previous
patch. Default value of multiple block is set as 4, in case if
the argument is not given by the user but the flag is provided.
This fixes part of #1709.

Signed-off-by: Mohana Datta Yelugoti <ymdatta.work@gmail.com>