kernel:sched: Give each CPU a sched_tick_timer
In SMP case, we use affinity to specify which cpu that a thread
belongs to. When we put all threads to one CPU, the rest of it do
nothing, IDLE_state cpu will delete the sched_tick_timer.
So it's necessary to give each cpu a timer in order to not disturb each
other